+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [v1.0.43] - 2025-04-18
+
+### Added
+- `--noGit` flag to `push` command, which deploys contents of working dir as-is, without special git rules (gitignore etc.)
+- checks to `push` command whether git is installed and correctly initialised when `--noGit` flag is not used
+
 ## [v1.0.42] - 2025-04-10
 
 ### Fixed